Added AndHow.java tests * Separated tests that require andhow from those that require anyhow-core
