+++ andhow-testing/andhow-core-system-tests/src/test/java/org/yarnandtail/andhow/property/StrPropTest.java
