+++ andhow-testing/andhow-core-system-tests/src/test/resources/org/yarnandtail/andhow/example/restclient/all.points.speced.properties
