+++ andhow-testing/andhow-core-system-tests/src/test/resources/org/yarnandtail/andhow/example/restclient/minimum.points.speced.properties
