+++ andhow-testing/andhow-core-system-tests/src/test/resources/org/yarnandtail/andhow/property/StrPropTest_validationGroupBad.properties
